requirements.txt
slackclient>=2.0.0
slackeventsapi
Flask>=1.1.2
covid-data-api
